requirements.txt
Django==2.2.1
django-ckeditor==5.8.0
django-cors-headers==3.0.2
django-js-asset==1.2.2
django-tinymce==2.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
